requirements.txt
pandas
numpy<2
scikit-learn
scikit-surprise
streamlit
joblib
matplotlib
jupyter
